refactor(index): slim the minhash index implementation and its tests
Split the module into mod (parameters, signatures, plugin), index (search,
cache, rebuilds) and builder (signing, external sort); share one text-array
iterator between the builder and the flat search; drop the statvfs check
before spills (the configured limit and the OS error suffice), the merge
document-count helper and the benchmark; and cut the tests to the behaviors
that matter: recall, spanning buckets, spilled builds, masks and caching,
update/remap/merge, deferred compaction, prewarm, the known-answer vectors
and the open-time checks.
