+++ rust/lance/Cargo.toml
@@ -235,10 +235,6 @@ name = "fts_search"
 harness = false
 
 [[bench]]
-name = "minhash_lsh"
-harness = false
-
-[[bench]]
 name = "vector_throughput"
 harness = false
